ipa: raspberrrypi: Add LIBCAMERA_NOTPARSE_EMBEDDED_DATA environment variable
Embedded data was not exported correctly for Quad cameras in older versions of Arducam.
The program always reports embedded data parsing errors, but it does not affect the
normal operation of the program. All we need is this environment variable for controlling
that the program no longer parses embedded data.